requirements.extra.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --constraint=/home/mjd/env-specs/ds-core/requirements.txt --output-file=requirements.extra.txt requirements.extra.in
#
attrs==25.4.0
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   ecmwf-datastores-client
beautifulsoup4==4.14.3
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   -r requirements.extra.in
cachetools==6.2.4
    # via google-auth
cdsapi==0.7.7
    # via -r requirements.extra.in
certifi==2025.11.12
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   pyogrio
    #   pyproj
    #   requests
charset-normalizer==3.4.4
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   requests
earthengine-api==1.7.4
    # via -r requirements.extra.in
ecmwf-datastores-client==0.4.1
    # via cdsapi
geopandas==1.1.2
    # via -r requirements.extra.in
google-api-core==2.28.1
    # via
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-storage
google-api-python-client==2.187.0
    # via earthengine-api
google-auth==2.45.0
    # via
    #   earthengine-api
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.3.0
    # via
    #   earthengine-api
    #   google-api-python-client
google-cloud-core==2.5.0
    # via google-cloud-storage
google-cloud-storage==3.7.0
    # via earthengine-api
google-crc32c==1.8.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.8.0
    # via google-cloud-storage
googleapis-common-protos==1.72.0
    # via google-api-core
httplib2==0.31.0
    # via
    #   earthengine-api
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.11
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   requests
lxml==6.0.2
    # via -r requirements.extra.in
multiurl==0.3.7
    # via ecmwf-datastores-client
numpy==2.3.5
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   geopandas
    #   pandas
    #   pyogrio
    #   shapely
packaging==25.0
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   geopandas
    #   pyogrio
pandas==2.3.3
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   -r requirements.extra.in
    #   geopandas
proto-plus==1.27.0
    # via google-api-core
protobuf==6.33.2
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pyogrio==0.12.1
    # via geopandas
pyparsing==3.3.1
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   httplib2
pyproj==3.7.2
    # via geopandas
python-dateutil==2.9.0.post0
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   multiurl
    #   pandas
python-dotenv==1.2.1
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   -r requirements.extra.in
pytz==2025.2
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   multiurl
    #   pandas
pyyaml==6.0.3
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   -r requirements.extra.in
requests==2.32.5
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   -r requirements.extra.in
    #   cdsapi
    #   earthengine-api
    #   ecmwf-datastores-client
    #   google-api-core
    #   google-cloud-storage
    #   multiurl
rsa==4.9.1
    # via google-auth
shapely==2.1.2
    # via geopandas
six==1.17.0
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   python-dateutil
soupsieve==2.8.1
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   beautifulsoup4
tqdm==4.67.1
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   -r requirements.extra.in
    #   cdsapi
    #   multiurl
typing-extensions==4.15.0
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   beautifulsoup4
    #   ecmwf-datastores-client
tzdata==2025.3
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.2
    # via
    #   -c /home/mjd/env-specs/ds-core/requirements.txt
    #   requests
